Combine two base_functions files for batch and gui render into one
